requirements.txt
pyyaml
pypdf
pandas
tiktoken
langchain
langgraph
langchain_community
langchain_openai
langchain_weaviate
langchain_huggingface
pymongo
sentence-transformers
weaviate-client
marker-pdf[full]
markdown-it-py
pyvips
einops
pyvips-binary
markdown
html5lib
ragas